AST-81788 - PoC - using gen-ai-prompts project
changes on project to make it able to be imported and used on other projects
