+++ internal/service/source/sourcehandler.go
@@ -1,4 +1,4 @@
-package sast_result_remediation
+package source
 
 import (
 	"bufio"
@@ -7,47 +7,52 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/Checkmarx/gen-ai-prompts/pkg/model"
 )
 
-type SourceAndError struct {
-	Source []string
-	Error  error
+type SourceHandlerImpl struct {
+	sourcePath string
 }
 
-func (pb *PromptBuilder) GetSourcesForResults(results []*Result) map[string]*SourceAndError {
+func NewSourceHandlerImpl(sourcePath string) *SourceHandlerImpl {
+	return &SourceHandlerImpl{sourcePath: sourcePath}
+}
+
+func (s *SourceHandlerImpl) GetSourcesForResults(results []*model.Result) map[string]*model.SourceAndError {
 	sourceFilenames := make(map[string]error)
 	for _, result := range results {
-		getFilenamesForResult(result, sourceFilenames)
+		s.getFilenamesForResult(result, sourceFilenames)
 	}
 
-	fileContents := make(map[string]*SourceAndError)
+	fileContents := make(map[string]*model.SourceAndError)
 	for filename, err := range sourceFilenames {
 		if err != nil {
-			fileContents[filename] = &SourceAndError{Source: nil, Error: err}
+			fileContents[filename] = &model.SourceAndError{Source: nil, Error: err}
 			continue
 		}
-		lines, err := pb.GetFileContents(filename)
+		lines, err := s.GetFileContents(filename)
 		if err != nil || lines == nil || len(lines) <= 1 {
 			if err == nil {
 				err = fmt.Errorf("file '%s' has irrelevant content", filename)
 			}
 			sourceFilenames[filename] = err
-			fileContents[filename] = &SourceAndError{Source: nil, Error: err}
+			fileContents[filename] = &model.SourceAndError{Source: nil, Error: err}
 		} else {
-			fileContents[filename] = &SourceAndError{Source: lines, Error: nil}
+			fileContents[filename] = &model.SourceAndError{Source: lines, Error: nil}
 		}
 	}
 
 	return fileContents
 }
 
-func (pb *PromptBuilder) GetSourcesForResult(result *Result) map[string]*SourceAndError {
-	results := []*Result{result}
-	return pb.GetSourcesForResults(results)
+func (s *SourceHandlerImpl) GetSourcesForResult(result *model.Result) map[string]*model.SourceAndError {
+	results := []*model.Result{result}
+	return s.GetSourcesForResults(results)
 }
 
-func (pb *PromptBuilder) GetFileContents(filename string) ([]string, error) {
-	sourceFilename := filepath.Join(pb.SourcePath, filename)
+func (s *SourceHandlerImpl) GetFileContents(filename string) ([]string, error) {
+	sourceFilename := filepath.Join(s.sourcePath, filename)
 	file, err := os.Open(sourceFilename)
 	if err != nil {
 		return nil, err
@@ -80,14 +85,14 @@ func (pb *PromptBuilder) GetFileContents(filename string) ([]string, error) {
 	return lines, nil
 }
 
-func getFilenamesForResult(result *Result, sourceFilenames map[string]error) {
+func (s *SourceHandlerImpl) getFilenamesForResult(result *model.Result, sourceFilenames map[string]error) {
 	for _, node := range result.Data.Nodes {
 		sourceFilename := strings.ReplaceAll(node.FileName, "\\", "/")
-		sourceFilenames[sourceFilename] = isResultRelevantForAnalysis(result, sourceFilename)
+		sourceFilenames[sourceFilename] = s.isResultRelevantForAnalysis(result, sourceFilename)
 	}
 }
 
-func isResultRelevantForAnalysis(result *Result, filename string) error {
+func (s *SourceHandlerImpl) isResultRelevantForAnalysis(result *model.Result, filename string) error {
 	var blacklistedExtensionsByLanguage = map[string][]string{
 		"JavaScript": []string{".min.js", "-min.js"},
 	}
