+++ pkg/service/prompt/build-prompt_test.go
@@ -1,14 +1,20 @@
-package sast_result_remediation
+package prompt
 
 import (
 	"fmt"
+	"reflect"
 	"testing"
+
+	"github.com/Checkmarx/gen-ai-prompts/internal/service/scanresults"
+	"github.com/Checkmarx/gen-ai-prompts/internal/service/source"
+	"github.com/Checkmarx/gen-ai-prompts/pkg/model"
+	"github.com/Checkmarx/gen-ai-prompts/pkg/utils"
 )
 
 const (
-	resultsFile     = "testdata/cx_result.json"
-	sourcePath      = "testdata/sources"
-	resultsListFile = "testdata/results_list.txt"
+	resultsFile     = "../../../test/cx_result.json"
+	sourcePath      = "../../../test/sources"
+	resultsListFile = "../../../test/results_list.txt"
 )
 
 const userPromptCodeSQLI = `  public AttackResult completed(@RequestParam String query) {//FILE: SqlInjectionLesson3.java:53//SAST Node #0 (input): query (ParamDecl)
@@ -54,6 +60,30 @@ const userPromptCode3 = `    $.get("challenge/8/votes/", function (votes) {//FIL
                 $("#nrOfVotes" + i).html(votes[i]);//SAST Node #1: votes (objectReference)//SAST Node #2 (output): html (MethodInvokeExpr)
 // method continues ...`
 
+const expectedOutputFormat = "<span style=\"color: regular;\">" + utils.BoldConfidence + "</span><span style=\"color: grey; font-style: italic;\"> " +
+	"A score between 0 (low) and 100 (high) indicating the degree of Confidence in the exploitability of this vulnerability in the context of your code. " +
+	"<br></span>%s<span style=\"color: regular;\">" + utils.BoldExplanation + "</span><span style=\"color: grey; font-style: italic;\"> " +
+	"An OpenAI generated description of the vulnerability. <br></span>%s<span style=\"color: " +
+	"regular;\">" + utils.BoldFix + "</span><span style=\"color: grey; font-style: italic;\"> " +
+	"A customized snippet, generated by OpenAI, that can be used to remediate the vulnerability in your code. <br></span>%s"
+
+const (
+	boldConfidence2  = utils.Bold + utils.Confidence + utils.Bold2
+	boldExplanation2 = utils.Bold + utils.Explanation + utils.Bold2
+	boldFix2         = utils.Bold + utils.Fix + utils.Bold2
+	boldConfidence3  = utils.Bold + utils.Confidence
+	boldExplanation3 = utils.Bold + utils.Explanation
+	boldFix3         = utils.Bold + utils.Fix
+	boldFix4         = utils.Bold + "PROPOSED  REMEDIATION" + utils.Bold2
+	plainConfidence  = utils.Confidence
+	plainExplanation = utils.Explanation
+	plainFix         = utils.Fix
+	header           = "##"
+	mixedFix         = "PROPOSED  REMEdiation"
+	explanationText  = " this is a short text explaining the score."
+	fixText          = " a fixed snippet"
+)
+
 func TestBuildPrompt(t *testing.T) {
 	type args struct {
 		resultsFile string
@@ -67,28 +97,30 @@ func TestBuildPrompt(t *testing.T) {
 		wantUser   string
 		wantErr    error
 	}{
-		{"TestBuildPromptHappy", args{"testdata/cx_result.json", "13588507", sourcePath},
-			systemPrompt, userPrompt("SQL_Injection", 89, "Java", userPromptCodeNodeLinesOnly), nil},
+		{"TestBuildPromptHappy", args{"../../../test/cx_result.json", "13588507", sourcePath},
+			utils.SystemPrompt, userPrompt("SQL_Injection", 89, "Java", userPromptCodeNodeLinesOnly), nil},
 		{"TestBuildPromptFileNotFound", args{"invalidFile", "13588507", sourcePath},
 			"", "", fmt.Errorf("error reading and parsing SAST results file 'invalidFile': 'open invalidFile: no such file or directory'")},
-		{"TestBuildPromptResultIdNotFound", args{"testdata/cx_result.json", "invalidResultId", sourcePath},
+		{"TestBuildPromptResultIdNotFound", args{"../../../test/cx_result.json", "invalidResultId", sourcePath},
 			"", "", fmt.Errorf("error getting result for result ID 'invalidResultId': 'result ID invalidResultId not found'")},
-		{"TestBuildPromptSourcesNotFound", args{"testdata/cx_result.json", "13588507", "invalidSources"},
-			systemPrompt, "", fmt.Errorf("error creating prompt for result ID '13588507': 'error reading source 'SqlInjectionLesson3.java': 'open invalidSources/SqlInjectionLesson3.java: no such file or directory''")},
+		{"TestBuildPromptSourcesNotFound", args{"../../../test/cx_result.json", "13588507", "invalidSources"},
+			utils.SystemPrompt, "", fmt.Errorf("error creating prompt for result ID '13588507': 'error reading source 'SqlInjectionLesson3.java': 'open invalidSources/SqlInjectionLesson3.java: no such file or directory''")},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			gotSystem, gotUser, gotError := BuildPrompt(tt.args.resultsFile, tt.args.resultId, tt.args.sourcePath)
-			if gotError != nil &&
-				gotError.Error() != tt.wantErr.Error() {
-				t.Errorf("BuildPrompt() error = %v, wantErr %v", gotError, tt.wantErr)
+			sastResults := scanresults.NewSastResultImpl()
+			sourceHandler := source.NewSourceHandlerImpl(tt.args.sourcePath)
+			pb := NewPromptBuilder(sastResults, sourceHandler, tt.args.resultsFile, tt.args.sourcePath, true)
+			promptResult := pb.BuildPromptForResultId(tt.args.resultId)
+			if promptResult.Error != nil && promptResult.Error.Error() != tt.wantErr.Error() {
+				t.Errorf("BuildPrompt() error = %v, wantErr %v", promptResult.Error, tt.wantErr)
 				return
 			}
-			if gotSystem != tt.wantSystem {
-				t.Errorf("BuildPrompt() gotSystem = %v, want %v", gotSystem, tt.wantSystem)
+			if promptResult.System != tt.wantSystem {
+				t.Errorf("BuildPrompt() gotSystem = %v, want %v", promptResult.System, tt.wantSystem)
 			}
-			if gotUser != tt.wantUser {
-				t.Errorf("BuildPrompt() gotUser = %v, want %v", gotUser, tt.wantUser)
+			if promptResult.User != tt.wantUser {
+				t.Errorf("BuildPrompt() gotUser = %v, want %v", promptResult.User, tt.wantUser)
 			}
 		})
 	}
@@ -108,20 +140,24 @@ func TestBuildPromptForResultIdNodeLinesOnly(t *testing.T) {
 		wantErr    error
 	}{
 		{"TestBuildPromptForResultIdAllLinesHappy", args{resultsFile, "13588507", sourcePath},
-			systemPrompt, userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI), nil},
+			utils.SystemPrompt, userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI), nil},
 		{"TestBuildPromptForResultIdAllLinesFileNotFound", args{"invalidFile", "13588507", sourcePath},
 			"", "", fmt.Errorf("error reading and parsing SAST results file 'invalidFile': 'open invalidFile: no such file or directory'")},
 		{"TestBuildPromptForResultIdAllLinesResultIdNotFound", args{resultsFile, "invalidResultId", sourcePath},
 			"", "", fmt.Errorf("error getting result for result ID 'invalidResultId': 'result ID invalidResultId not found'")},
 		{"TestBuildPromptForResultIdAllLinesSourcesNotFound", args{resultsFile, "13588507", "invalidSources"},
-			systemPrompt, "", fmt.Errorf("error creating prompt for result ID '13588507': 'error getting method 'completed': 'error reading source 'SqlInjectionLesson3.java': open invalidSources/SqlInjectionLesson3.java: no such file or directory''")},
+			utils.SystemPrompt, "", fmt.Errorf("error creating prompt for result ID '13588507': 'error getting method 'completed': 'error reading source 'SqlInjectionLesson3.java': open invalidSources/SqlInjectionLesson3.java: no such file or directory''")},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			sr := scanresults.NewSastResultImpl()
+			sh := source.NewSourceHandlerImpl(tt.args.sourcePath)
 			pb := &PromptBuilder{
 				ResultsFile:   tt.args.resultsFile,
 				SourcePath:    tt.args.sourcePath,
 				NodeLinesOnly: false,
+				sastResults:   sr,
+				sourceHandler: sh,
 			}
 			got := pb.BuildPromptForResultId(tt.args.resultId)
 			if got.Error != nil &&
@@ -150,17 +186,17 @@ func TestBuildPromptsForLanguageAndQuery(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want []SastResultPrompt
+		want []model.SastResultPrompt
 	}{
 		{"TestBuildPromptsForJavaSqlInjection",
 			args{resultsFile: resultsFile, sourcePath: sourcePath, language: "java", query: "sql_Injection", resultId: ""},
-			[]SastResultPrompt{
+			[]model.SastResultPrompt{
 				{
 					ResultsFile: resultsFile,
 					SourcePath:  sourcePath,
 					Language:    "Java",
 					Query:       "SQL_Injection",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI),
 					Error:       nil,
 				},
@@ -169,7 +205,7 @@ func TestBuildPromptsForLanguageAndQuery(t *testing.T) {
 					SourcePath:  sourcePath,
 					Language:    "Java",
 					Query:       "SQL_Injection",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI2),
 					Error:       nil,
 				},
@@ -177,13 +213,13 @@ func TestBuildPromptsForLanguageAndQuery(t *testing.T) {
 		},
 		{"TestBuildPromptsForJavaScript",
 			args{resultsFile: resultsFile, sourcePath: sourcePath, language: "javascript", query: "*", resultId: ""},
-			[]SastResultPrompt{
+			[]model.SastResultPrompt{
 				{
 					ResultsFile: resultsFile,
 					SourcePath:  sourcePath,
 					Language:    "JavaScript",
 					Query:       "Client_DOM_Stored_XSS",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("Client_DOM_Stored_XSS", 79, "JavaScript", userPromptCode3),
 					Error:       nil,
 				},
@@ -192,7 +228,7 @@ func TestBuildPromptsForLanguageAndQuery(t *testing.T) {
 					SourcePath:  sourcePath,
 					Language:    "JavaScript",
 					Query:       "Client_DOM_Open_Redirect",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        "",
 					Error:       fmt.Errorf("error creating prompt for result ID '13588372': 'error getting method 'Cxd39430bd': 'error reading source '/backbone-min.js': blacklisted extension: -min.js''"),
 				},
@@ -201,10 +237,14 @@ func TestBuildPromptsForLanguageAndQuery(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			sr := scanresults.NewSastResultImpl()
+			sh := source.NewSourceHandlerImpl(tt.args.sourcePath)
 			pb := &PromptBuilder{
 				ResultsFile:   tt.args.resultsFile,
 				SourcePath:    tt.args.sourcePath,
 				NodeLinesOnly: false,
+				sastResults:   sr,
+				sourceHandler: sh,
 			}
 			got := pb.BuildPromptsForLanguageAndQuery(tt.args.language, tt.args.query)
 			if len(got) != len(tt.want) {
@@ -237,17 +277,17 @@ func TestBuildPromptsForSeverity(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want []SastResultPrompt
+		want []model.SastResultPrompt
 	}{
 		{"TestBuildPromptsForHigh",
 			args{resultsFile: resultsFile, sourcePath: sourcePath, severity: "high"},
-			[]SastResultPrompt{
+			[]model.SastResultPrompt{
 				{
 					ResultsFile: resultsFile,
 					SourcePath:  sourcePath,
 					Language:    "Java",
 					Query:       "SQL_Injection",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI),
 					Error:       nil,
 				},
@@ -256,7 +296,7 @@ func TestBuildPromptsForSeverity(t *testing.T) {
 					SourcePath:  sourcePath,
 					Language:    "Java",
 					Query:       "Client_DOM_Stored_XSS",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("Client_DOM_Stored_XSS", 79, "JavaScript", userPromptCode3),
 					Error:       nil,
 				},
@@ -264,13 +304,13 @@ func TestBuildPromptsForSeverity(t *testing.T) {
 		},
 		{"TestBuildPromptsForLow",
 			args{resultsFile: resultsFile, sourcePath: sourcePath, severity: "Low"},
-			[]SastResultPrompt{
+			[]model.SastResultPrompt{
 				{
 					ResultsFile: resultsFile,
 					SourcePath:  sourcePath,
 					Language:    "JavaScript",
 					Query:       "Client_DOM_Open_Redirect",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        "",
 					Error:       fmt.Errorf("error creating prompt for result ID '13588372': 'error getting method 'Cxd39430bd': 'error reading source '/backbone-min.js': blacklisted extension: -min.js''"),
 				},
@@ -279,7 +319,7 @@ func TestBuildPromptsForSeverity(t *testing.T) {
 					SourcePath:  sourcePath,
 					Language:    "Java",
 					Query:       "SQL_Injection",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI2),
 					Error:       nil,
 				},
@@ -287,7 +327,7 @@ func TestBuildPromptsForSeverity(t *testing.T) {
 		},
 		{"TestBuildPromptsForOther",
 			args{resultsFile: resultsFile, sourcePath: sourcePath, severity: "Junk"},
-			[]SastResultPrompt{
+			[]model.SastResultPrompt{
 				{
 					ResultsFile: resultsFile,
 					SourcePath:  sourcePath,
@@ -302,10 +342,14 @@ func TestBuildPromptsForSeverity(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			sr := scanresults.NewSastResultImpl()
+			sh := source.NewSourceHandlerImpl(tt.args.sourcePath)
 			pb := &PromptBuilder{
 				ResultsFile:   tt.args.resultsFile,
 				SourcePath:    tt.args.sourcePath,
 				NodeLinesOnly: false,
+				sastResults:   sr,
+				sourceHandler: sh,
 			}
 			got := pb.BuildPromptsForSeverity(tt.args.severity)
 			if len(got) != len(tt.want) {
@@ -342,17 +386,17 @@ func TestBuildPromptsForResultsList(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want []SastResultPrompt
+		want []model.SastResultPrompt
 	}{
 		{"TestBuildPromptsForResultsList",
 			args{resultsFile: resultsFile, sourcePath: sourcePath, resultsListFile: resultsListFile},
-			[]SastResultPrompt{
+			[]model.SastResultPrompt{
 				{
 					ResultsFile: resultsFile,
 					SourcePath:  sourcePath,
 					Language:    "Java",
 					Query:       "SQL_Injection",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI),
 					Error:       nil,
 				},
@@ -361,7 +405,7 @@ func TestBuildPromptsForResultsList(t *testing.T) {
 					SourcePath:  sourcePath,
 					Language:    "Java",
 					Query:       "SQL_Injection",
-					System:      systemPrompt,
+					System:      utils.SystemPrompt,
 					User:        userPrompt("SQL_Injection", 89, "Java", userPromptCodeSQLI2),
 					Error:       nil,
 				},
@@ -370,10 +414,14 @@ func TestBuildPromptsForResultsList(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			sr := scanresults.NewSastResultImpl()
+			sh := source.NewSourceHandlerImpl(tt.args.sourcePath)
 			pb := &PromptBuilder{
 				ResultsFile:   tt.args.resultsFile,
 				SourcePath:    tt.args.sourcePath,
 				NodeLinesOnly: false,
+				sastResults:   sr,
+				sourceHandler: sh,
 			}
 			got := pb.BuildPromptsForResultsListFile(tt.args.resultsListFile)
 			if len(got) != len(tt.want) {
@@ -398,7 +446,7 @@ func TestBuildPromptsForResultsList(t *testing.T) {
 }
 
 func userPrompt(query string, cwe int, language string, code string) string {
-	return fmt.Sprintf(userPromptTemplate, query, cwe, language, code)
+	return fmt.Sprintf(utils.UserPromptTemplate, query, cwe, language, code)
 }
 
 const codeMissingNode = `    public String getRealPath(String path) {//FILE: JspCServletContext.java:296
@@ -704,48 +752,52 @@ func TestBuildPromptForResults(t *testing.T) {
 		wantErr    error
 	}{
 		{"TestBuildPromptForResultMissingNode",
-			args{"testdata/sast_result_missing-node.json", "c1CrCRw4/3/A6q+6zwIhShQIe1I=", sourcePath, false},
-			systemPrompt, userPrompt("Stored_XSS", 79, "Java", codeMissingNode), nil},
+			args{"../../../test/sast_result_missing-node.json", "c1CrCRw4/3/A6q+6zwIhShQIe1I=", sourcePath, false},
+			utils.SystemPrompt, userPrompt("Stored_XSS", 79, "Java", codeMissingNode), nil},
 		{"TestBuildPromptTwoSimilarResults1",
-			args{"testdata/two_similar_results.json", "13893625", sourcePath, false},
-			systemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults1), nil},
+			args{"../../../test/two_similar_results.json", "13893625", sourcePath, false},
+			utils.SystemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults1), nil},
 		{"TestBuildPromptTwoSimilarResults2",
-			args{"testdata/two_similar_results.json", "13893626", sourcePath, false},
-			systemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults2), nil},
+			args{"../../../test/two_similar_results.json", "13893626", sourcePath, false},
+			utils.SystemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults2), nil},
 		{"TestBuildPromptJspResult",
-			args{"testdata/jsp_result.json", "vuKUhCJ5drCeY6IDB//eBu8wvkk=", sourcePath, false},
-			systemPrompt, userPrompt("LDAP_Injection", 90, "Java", jspCode), nil},
+			args{"../../../test/jsp_result.json", "vuKUhCJ5drCeY6IDB//eBu8wvkk=", sourcePath, false},
+			utils.SystemPrompt, userPrompt("LDAP_Injection", 90, "Java", jspCode), nil},
 		{"TestBuildPromptJspAndJavaResult",
-			args{"testdata/jsp_and_java_result.json", "XrM9Lk/bjJHxLOcn4XETGHJ1ko0=", sourcePath, false},
-			systemPrompt, userPrompt("Reflected_XSS_All_Clients", 79, "Java", jspAndJavaCode), nil},
+			args{"../../../test/jsp_and_java_result.json", "XrM9Lk/bjJHxLOcn4XETGHJ1ko0=", sourcePath, false},
+			utils.SystemPrompt, userPrompt("Reflected_XSS_All_Clients", 79, "Java", jspAndJavaCode), nil},
 		{"TestBuildPromptTwoMethodsWithSameName",
-			args{"testdata/two_methods_with_same_name.json", "5LcQqrUhDTZWEVCBWwMWGhSm+00=", sourcePath, false},
-			systemPrompt, userPrompt("Input_Path_Not_Canonicalized", 73, "Java", codeTwoMethodsWithSameName), nil},
+			args{"../../../test/two_methods_with_same_name.json", "5LcQqrUhDTZWEVCBWwMWGhSm+00=", sourcePath, false},
+			utils.SystemPrompt, userPrompt("Input_Path_Not_Canonicalized", 73, "Java", codeTwoMethodsWithSameName), nil},
 		{"TestBuildPromptForResultMissingNodeNLO",
-			args{"testdata/sast_result_missing-node.json", "c1CrCRw4/3/A6q+6zwIhShQIe1I=", sourcePath, true},
-			systemPrompt, userPrompt("Stored_XSS", 79, "Java", codeMissingNodeNLO), nil},
+			args{"../../../test/sast_result_missing-node.json", "c1CrCRw4/3/A6q+6zwIhShQIe1I=", sourcePath, true},
+			utils.SystemPrompt, userPrompt("Stored_XSS", 79, "Java", codeMissingNodeNLO), nil},
 		{"TestBuildPromptTwoSimilarResults1NLO",
-			args{"testdata/two_similar_results.json", "13893625", sourcePath, true},
-			systemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults1NLO), nil},
+			args{"../../../test/two_similar_results.json", "13893625", sourcePath, true},
+			utils.SystemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults1NLO), nil},
 		{"TestBuildPromptTwoSimilarResults2NLO",
-			args{"testdata/two_similar_results.json", "13893626", sourcePath, true},
-			systemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults2NLO), nil},
+			args{"../../../test/two_similar_results.json", "13893626", sourcePath, true},
+			utils.SystemPrompt, userPrompt("Log_Forging", 117, "Java", codeTwoSimilarResults2NLO), nil},
 		{"TestBuildPromptJspResultNLO",
-			args{"testdata/jsp_result.json", "vuKUhCJ5drCeY6IDB//eBu8wvkk=", sourcePath, true},
-			systemPrompt, userPrompt("LDAP_Injection", 90, "Java", jspCodeNLO), nil},
+			args{"../../../test/jsp_result.json", "vuKUhCJ5drCeY6IDB//eBu8wvkk=", sourcePath, true},
+			utils.SystemPrompt, userPrompt("LDAP_Injection", 90, "Java", jspCodeNLO), nil},
 		{"TestBuildPromptJspAndJavaResultNLO",
-			args{"testdata/jsp_and_java_result.json", "XrM9Lk/bjJHxLOcn4XETGHJ1ko0=", sourcePath, true},
-			systemPrompt, userPrompt("Reflected_XSS_All_Clients", 79, "Java", jspAndJavaCodeNLO), nil},
+			args{"../../../test/jsp_and_java_result.json", "XrM9Lk/bjJHxLOcn4XETGHJ1ko0=", sourcePath, true},
+			utils.SystemPrompt, userPrompt("Reflected_XSS_All_Clients", 79, "Java", jspAndJavaCodeNLO), nil},
 		{"TestBuildPromptTwoMethodsWithSameNameNLO",
-			args{"testdata/two_methods_with_same_name.json", "5LcQqrUhDTZWEVCBWwMWGhSm+00=", sourcePath, true},
-			systemPrompt, userPrompt("Input_Path_Not_Canonicalized", 73, "Java", codeTwoMethodsWithSameNameNLO), nil},
+			args{"../../../test/two_methods_with_same_name.json", "5LcQqrUhDTZWEVCBWwMWGhSm+00=", sourcePath, true},
+			utils.SystemPrompt, userPrompt("Input_Path_Not_Canonicalized", 73, "Java", codeTwoMethodsWithSameNameNLO), nil},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			sr := scanresults.NewSastResultImpl()
+			sh := source.NewSourceHandlerImpl(tt.args.sourcePath)
 			pb := &PromptBuilder{
 				ResultsFile:   tt.args.resultsFile,
 				SourcePath:    tt.args.sourcePath,
 				NodeLinesOnly: tt.args.nodeLinesOnly,
+				sastResults:   sr,
+				sourceHandler: sh,
 			}
 			got := pb.BuildPromptForResultId(tt.args.resultId)
 			if got.Error != nil &&
@@ -761,4 +813,120 @@ func TestBuildPromptForResults(t *testing.T) {
 			}
 		})
 	}
+}
+
+func getExpectedOutput(confidenceNumber, explanationText, fixText string) string {
+	return fmt.Sprintf(expectedOutputFormat, confidenceNumber, explanationText, fixText)
+}
+
+func TestAddDescriptionForIdentifiers(t *testing.T) {
+	input := utils.BoldConfidence + " 35 " + utils.BoldExplanation + explanationText + utils.BoldFix + fixText
+	expected := getExpectedOutput(" 35 ", explanationText, fixText)
+	output := getActual(input, t)
+	if output[len(output)-1] != expected {
+		t.Errorf("Expected %q, but got %q", expected, output[len(output)-1])
+	}
+}
+
+func TestAddNewlinesIfNecessarySomeNewlines(t *testing.T) {
+	input := utils.BoldConfidence + " 35 " + utils.BoldExplanation + explanationText + "\n" + utils.BoldFix + fixText
+	expected := getExpectedOutput(" 35 ", explanationText+"\n", fixText)
+
+	output := getActual(input, t)
+
+	if output[len(output)-1] != expected {
+		t.Errorf("Expected %q, but got %q", expected, output[len(output)-1])
+	}
+}
+
+func TestAddNewlinesIfNecessaryAllNewlines(t *testing.T) {
+	input := utils.BoldConfidence + " 35\n " + utils.BoldExplanation + explanationText + "\n" + utils.BoldFix + fixText
+	expected := getExpectedOutput(" 35\n ", explanationText+"\n", fixText)
+
+	output := getActual(input, t)
+
+	if output[len(output)-1] != expected {
+		t.Errorf("Expected %q, but got %q", expected, output[len(output)-1])
+	}
+}
+
+func TestParseResponse(t *testing.T) {
+	introText := "this is some introductory text"
+	goodConfidenceText := " 35\n"
+	goodConfidenceText3 := " 35**\n"
+	badConfidenceText := "0\nfailed0"
+	badConfidenceText2 := "N/A"
+	confidenceValue := 35
+	explanationText := explanationText + "\n"
+
+	tests := []struct {
+		name     string
+		input    string
+		expected *model.ParsedResponse
+		err      error
+	}{
+		{"TestParseResponseHappy", introText + utils.BoldConfidence + goodConfidenceText + utils.BoldExplanation + explanationText + utils.BoldFix + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: confidenceValue, Explanation: explanationText, Fix: fixText}, nil},
+		{"TestParseResponseHappy2", introText + boldConfidence2 + goodConfidenceText + utils.BoldExplanation + explanationText + utils.BoldFix + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: confidenceValue, Explanation: explanationText, Fix: fixText}, nil},
+		{"TestParseResponseHappy3", introText + boldConfidence3 + goodConfidenceText3 + utils.BoldExplanation + explanationText + utils.BoldFix + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: confidenceValue, Explanation: explanationText, Fix: fixText}, nil},
+		{"TestParseResponseHappy4", introText + utils.BoldConfidence + goodConfidenceText + boldExplanation2 + explanationText + boldFix2 + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: confidenceValue, Explanation: explanationText, Fix: fixText}, nil},
+		{"TestParseResponseHappy5", introText + utils.BoldConfidence + goodConfidenceText + boldExplanation3 + explanationText + boldFix3 + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: confidenceValue, Explanation: explanationText, Fix: fixText}, nil},
+		{"TestParseResponseHappy6", introText + utils.BoldConfidence + goodConfidenceText + boldExplanation3 + explanationText + boldFix4 + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: confidenceValue, Explanation: explanationText, Fix: fixText}, nil},
+		{"TestParseResponseHappy7", introText + header + plainConfidence + goodConfidenceText + header + plainExplanation + explanationText + header + plainFix + fixText,
+			&model.ParsedResponse{Introduction: introText + header, Confidence: confidenceValue, Explanation: explanationText + header, Fix: fixText}, nil},
+		{"TestParseResponseHappy8", introText + utils.BoldConfidence + goodConfidenceText + utils.BoldExplanation + explanationText + mixedFix + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: confidenceValue, Explanation: explanationText, Fix: fixText}, nil},
+		{"TestParseResponseNoConfidence", introText + goodConfidenceText + utils.BoldExplanation + explanationText + utils.BoldFix + fixText,
+			&model.ParsedResponse{Introduction: "", Confidence: 0, Explanation: "", Fix: ""},
+			fmt.Errorf("confidence not found in response")},
+		{"TestParseResponseNoExplanation", introText + utils.BoldConfidence + goodConfidenceText + explanationText + utils.BoldFix + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: 0, Explanation: "", Fix: ""},
+			fmt.Errorf("explanation not found in response")},
+		{"TestParseResponseNoFix", introText + utils.BoldConfidence + goodConfidenceText + utils.BoldExplanation + explanationText + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: 0, Explanation: "", Fix: ""},
+			fmt.Errorf("fix not found in response")},
+		{"TestParseResponseBadConfidenceValue", introText + utils.BoldConfidence + badConfidenceText + utils.BoldExplanation + explanationText + utils.BoldFix + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: 0, Explanation: explanationText, Fix: fixText},
+			fmt.Errorf("error converting confidence text to integer value: EOF")},
+		{"TestParseResponseBadConfidenceValue", introText + utils.BoldConfidence + badConfidenceText2 + utils.BoldExplanation + explanationText + utils.BoldFix + fixText,
+			&model.ParsedResponse{Introduction: introText, Confidence: 0, Explanation: explanationText, Fix: fixText},
+			fmt.Errorf("error converting confidence text to integer value: EOF")},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			sastResults := scanresults.NewSastResultImpl()
+			sourceHandler := source.NewSourceHandlerImpl("")
+			promptBuilder := NewPromptBuilder(sastResults, sourceHandler, "", "", true)
+			actual, err := promptBuilder.ParseResponse(tt.input)
+			if err != nil &&
+				(err.Error() != tt.err.Error()) {
+				t.Errorf("ParseResponse() gotErr = \n%v\nwantErr \n%v\n", err, tt.err)
+			}
+
+			if !reflect.DeepEqual(*actual, *tt.expected) {
+				t.Errorf("Expected %q, but got %q", *tt.expected, *actual)
+			}
+		})
+	}
+}
+
+func getActual(input string, t *testing.T) []string {
+	someText := "some text"
+	response := []string{someText, someText, input}
+	sastResults := scanresults.NewSastResultImpl()
+	sourceHandler := source.NewSourceHandlerImpl("")
+	promptBuilder := NewPromptBuilder(sastResults, sourceHandler, "", "", true)
+	output := promptBuilder.AddDescriptionForIdentifier(response)
+	for i := 0; i < len(output)-1; i++ {
+		if output[i] != response[i] {
+			t.Errorf("All strings except last expected to stay the same")
+		}
+	}
+	return output
 }
