+++ test/sources/challenge8.js
